Fix sampling error and increase to 16k hz.
The data from the sound card is in bytes. 2 bytes per integer (short).

It used to be converted directly to integer as if it were 1 byte per integer.

Values were low and every second value was about zero. This I could see after
plotting the data. Looked suspicious, so I investigated and what-a-surprise, my
code had a bug!